bringing handlers closer together
